Closes #13: Add a confirmation dialog on entry removal

+++ public/app/js/controllers.js
@@ -35,113 +35,105 @@ function showEntryDialog(entry, $modal) {
 angular.module('KeyRing.controllers', [])
 
 .controller('MainCtrl', ['$rootScope', '$q', '$modal', 'categoriesDAO', 'entriesDAO', 'cryptoSRV', function($rootScope, $q, $modal, categoriesDAO, entriesDAO, cryptoSRV) {
-   $rootScope.categories = [];
-   $rootScope.entries = [];
+  $rootScope.categories = [];
+  $rootScope.entries = [];
 
-   this.showCategories = function() {
-      var modalInstance = $modal.open({
-         templateUrl: 'partials/categories.html',
-         controller: 'CategoriesCtrl as Ctrl'
-      });
+  this.showCategories = function() {
+    var modalInstance = $modal.open({
+      templateUrl: 'partials/categories.html',
+      controller: 'CategoriesCtrl as Ctrl'
+    });
 
-      modalInstance.result.then(null, function(categories) {
-         // TODO
-      });
-   };
+    modalInstance.result.then(null, function(categories) {
+      // TODO
+    });
+  };
 
-   this.addNewEntry = function() {
-      showEntryDialog(null, $modal);
-   };
+  this.addNewEntry = function() {
+    showEntryDialog(null, $modal);
+  };
 
-   var loadCategories = function() {
-      return $q.when(categoriesDAO.load()).then(function(payload) { $rootScope.categories = payload.data.data; });
-   };
+  var loadCategories = function() {
+    return $q.when(categoriesDAO.load()).then(function(payload) { $rootScope.categories = payload.data.data; });
+  };
 
-   var loadEntries = function() {
-      return $q.when(entriesDAO.load()).then(function(entries) { 
-         $rootScope.entries = entries.data.data; 
-         $rootScope.checkLock();
-      });
-   };
+  var loadEntries = function() {
+    return $q.when(entriesDAO.load()).then(function(entries) { 
+      $rootScope.entries = entries.data.data; 
+      $rootScope.checkLock();
+    });
+  };
 
-   $rootScope.checkLock = function() {
-      $rootScope.entries = $rootScope.entries.map(function(entry) {
-         if ($rootScope.masterPassword) {
-            entry.clearUser = cryptoSRV.decrypt(entry.user, $rootScope.masterPassword);
-            entry.clearPassword = cryptoSRV.decrypt(entry.password, $rootScope.masterPassword);
-         } else {
-            delete entry.clearUser;
-            delete entry.clearPassword;
-         }
-         return entry;
-      });
-   };
-
-   $rootScope.reloadData = function() {
-      loadCategories().then(loadEntries);
-   };
-
-   $rootScope.reloadData();
-
-}])
-
-.controller('MasterPasswordCtrl', ['$rootScope', '$scope', '$modal', function($rootScope, $scope, $modal) {
-   this.show = function() {
-      var modalInstance = $modal.open({
-         templateUrl: 'partials/masterPassword.html',
-         controller: 'ChangeMasterPasswordCtrl as Ctrl',
-         resolve: {
-            masterPassword: function() {
-               return "";
-            }
-         }
-      });
-
-      modalInstance.result.then(function(password) {
-         $rootScope.masterPassword = password;
-         $rootScope.checkLock();
-      }, function() {
-         $rootScope.masterPassword = null;
-         $rootScope.checkLock();
-      });
-   };
-
-}])
-
-.controller('ChangeMasterPasswordCtrl', ['$scope', '$modalInstance', function($scope, $modalInstance) {
-   this.accept = function(keyEvent) {
-      if (typeof(keyEvent) == "undefined" || keyEvent.which == 13) {
-         $modalInstance.close($scope.masterPassword);
+  $rootScope.checkLock = function() {
+    $rootScope.entries = $rootScope.entries.map(function(entry) {
+      if ($rootScope.masterPassword) {
+        entry.clearUser = cryptoSRV.decrypt(entry.user, $rootScope.masterPassword);
+        entry.clearPassword = cryptoSRV.decrypt(entry.password, $rootScope.masterPassword);
+      } else {
+        delete entry.clearUser;
+        delete entry.clearPassword;
       }
-   };
+      return entry;
+    });
+  };
 
-   this.cancel = function() {
-      $modalInstance.close();
+  $rootScope.reloadData = function() {
+    loadCategories().then(loadEntries);
+  };
+
+  $rootScope.reloadData();
+
+}])
+
+.controller('MasterPasswordCtrl', ['$rootScope', '$scope', function($rootScope, $scope) {
+   this.show = function() {
+
+     bootbox.prompt({
+       title: "Master password",
+       className: 'info',
+       inputType: 'password',
+       callback: function(password) {
+         if (password) {
+           $rootScope.masterPassword = password;
+         } else {
+           $rootScope.masterPassword = null;
+         }
+        $rootScope.checkLock();
+       }
+     });
    };
 
 }])
 
 .controller('EntriesCtrl', ['$rootScope', '$scope', '$q', '$modal', '$log', 'entriesDAO', function($rootScope, $scope, $q, $modal, $log, entriesDAO) {
-   this.entriesByCategory = function(category) {
-      return $rootScope.entries.filter(function(entry) {
-         return entry.category === category.name;
-      });
-   };
+  this.entriesByCategory = function(category) {
+    return $rootScope.entries.filter(function(entry) {
+      return entry.category === category.name;
+    });
+  };
 
-   this.editEntry = function(entry) {
-      showEntryDialog(entry, $modal);
-   };
+  this.editEntry = function(entry) {
+    showEntryDialog(entry, $modal);
+  };
 
-   this.deleteEntry = function(entry) {
-      $q.when(entriesDAO.remove(entry)).then(function() { 
-         var index = $rootScope.entries.indexOf(entry);
-         $rootScope.entries.splice(index, 1);
-      }, function(err) {
-         $log.error(err);
-         alert("Error deleting entry: " + err.statusText);
-      });
-
-   }
+  this.deleteEntry = function(entry) {
+    bootbox.confirm({
+      title: "Confirm deletion",
+      message: "Are you sure you want to remove this entry?",
+      className: 'warning',
+      callback: function(result) {
+        if (result) {
+          $q.when(entriesDAO.remove(entry)).then(function() { 
+            var index = $rootScope.entries.indexOf(entry);
+            $rootScope.entries.splice(index, 1);
+          }, function(err) {
+            $log.error(err);
+            alert("Error deleting entry: " + err.statusText);
+          });
+        }
+      }
+    });
+  }
 }])
 
 .controller('EditEntryCtrl', ['$rootScope', '$scope', '$q', '$modalInstance', '$log', 'entriesDAO', 'cryptoSRV', 'entry', function($rootScope, $scope, $q, $modalInstance, $log, entriesDAO, cryptoSRV, entry) {
@@ -187,44 +179,53 @@ angular.module('KeyRing.controllers', [])
 }])
 
 .controller('CategoriesCtrl', ['$rootScope', '$scope', '$q', '$modalInstance', '$modal', '$log', 'categoriesDAO', function($rootScope, $scope, $q, $modalInstance, $modal, $log, categoriesDAO) {
-   this.remove = function(category) {
-      $q.when(categoriesDAO.remove(category)).then(function(error) {
-         //TODO: handle error
-         $rootScope.categories = $rootScope.categories.filter(function(cat) {
-            return cat.name != category.name;
-         });
-      });
-   };
-
-   this.add = function() {
-      var modalInstance = $modal.open({
-         templateUrl: 'partials/editCategory.html',
-         controller: 'EditCategoryCtrl as Ctrl',
-         //scope: $scope,
-         resolve: {
-            category: function() {
-               return { name: null };
-            }
-         }
-      });
-
-      modalInstance.result.then(function(newCategory) {
-         if (newCategory) {
-            $q.when(categoriesDAO.create(newCategory)).then(function(category) { 
-               $rootScope.reloadData();
-            }, function(err) {
-               $log.error(err);
-               alert("Error creating category: " + err.statusText);
+  this.remove = function(category) {
+    bootbox.confirm({
+      title: "Confirm deletion",
+      message: "Are you sure you want to remove this category?",
+      className: 'warning',
+      callback: function(result) {
+        if (result) {
+          $q.when(categoriesDAO.remove(category)).then(function(error) {
+            //TODO: handle error
+            $rootScope.categories = $rootScope.categories.filter(function(cat) {
+              return cat.name != category.name;
             });
-         }
-      }, function() {
-         // dialog closed without saving
-      });
-   };
+          });
+        }
+      }
+    });
+  };
 
-   this.cancel = function() {
-      $modalInstance.close();
-   };
+  this.add = function() {
+    var modalInstance = $modal.open({
+      templateUrl: 'partials/editCategory.html',
+      controller: 'EditCategoryCtrl as Ctrl',
+      //scope: $scope,
+      resolve: {
+        category: function() {
+          return { name: null };
+        }
+      }
+    });
+
+    modalInstance.result.then(function(newCategory) {
+      if (newCategory) {
+        $q.when(categoriesDAO.create(newCategory)).then(function(category) { 
+          $rootScope.reloadData();
+        }, function(err) {
+          $log.error(err);
+          alert("Error creating category: " + err.statusText);
+        });
+      }
+    }, function() {
+      // dialog closed without saving
+    });
+  };
+
+  this.cancel = function() {
+    $modalInstance.close();
+  };
 }])
 
 .controller('EditCategoryCtrl', ['$scope', '$modalInstance', 'category', function($scope, $modalInstance, category) {
